Update extension webpack configs
